Rename python package to tiledb.bioimg
