+++ .github/workflows/ci.yml
@@ -45,7 +45,7 @@ jobs:
         AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
         AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
       run: |
-        pytest -v --cov=tiledbimg --cov-report=term-missing --durations=0 tests/ > coverage.txt
+        pytest -v --cov=tiledb --cov-report=term-missing --durations=0 tests/ > coverage.txt
         cat coverage.txt
         TEST_COVERAGE="$(grep '^TOTAL' coverage.txt | awk -v N=4 '{print $N}')"
         echo "COVERAGE=$TEST_COVERAGE" >> $GITHUB_OUTPUT
