+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = tiledbimg
+name = tiledb-bioimg
 description = Package supports all bio-imaging functionality provided by TileDB
 author = TileDB, Inc.
 author_email = help@tiledb.io
@@ -7,7 +7,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 license = MIT
 license_file = LICENSE
-keywords = tiledb, tiledbimg , bioimaging
+keywords = tiledb, tiledb-bioimg , bioimaging
 url = https://github.com/TileDB-Inc/TileDB-Bioimaging
 platform = any
 project_urls =
