+++ tests/integration/converters/test_ome_tiff.py
@@ -4,8 +4,8 @@ import tifffile
 import tiledb
 
 from tests import get_path, get_schema
-from tiledbimg.converters.ome_tiff import OMETiffConverter
-from tiledbimg.openslide import TileDBOpenSlide
+from tiledb.bioimg.converters.ome_tiff import OMETiffConverter
+from tiledb.bioimg.openslide import TileDBOpenSlide
 
 
 def test_ome_tiff_converter(tmp_path):
