+++ tiledb/bioimg/converters/openslide.py
