+++ tiledb/bioimg/openslide.py
@@ -4,6 +4,7 @@ from operator import itemgetter
 from typing import Sequence, Tuple
 
 import numpy as np
+
 import tiledb
